move components into correct folders
