+++ tv-shows-app/src/app/(show_sidebar)/top-rated/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { AuthRedirect } from "@/components/shared/AuthRedirect/AuthRedirect";
-import ShowListContainer from "@/components/shared/ShowListContainer/ShowListContainer";
+import ShowListContainer from "@/components/features/ShowListContainer/ShowListContainer";
 
 export default function TopRated() {
   return (
