+++ tv-shows-app/src/components/features/ShowCard/ShowCard.tsx
