+++ tv-shows-app/src/components/features/ShowListContainer/ShowListContainer.tsx
@@ -37,7 +37,4 @@ export default function ShowListContainer({
   }
 
   return <ShowList showList={data?.shows || []} />;
-  {
-    shows: [];
-  }
 }
