+++ tv-shows-app/src/components/features/review/ReviewList/ReviewList.test.tsx
@@ -1,4 +1,4 @@
-import ShowList from "@/components/shared/ShowList/ShowList";
+import ShowList from "@/components/features/ShowList/ShowList";
 import { IReviewList } from "@/typings/review";
 import { render } from "@testing-library/react";
 import ReviewList from "./ReviewList";
